requirements.txt
build==1.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptg==0.5.2
cryptography==46.0.3
dhanhq==2.0.2
idna==3.11
isort==7.0.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
polars==1.35.2
polars-runtime-32==1.35.2
pyaes==1.6.1
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pyOpenSSL==25.3.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
ruff==0.14.8
six==1.17.0
TA-Lib==0.6.8
Telethon==1.42.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
